Move packages from pillar to map.jinja
Previously packages were stored in pillar variables.  This could have a
few potential problems: first, it could be undefined and fail silently.
Second, it did not conform to community standards for salt formulas.
Third, it obscured what packages variables were needed to be overridden
for a package. Finally, package variable names are not generally used in
multiple formula (at time of writing this did does not happen).

Now the packages are listed as variables in a file called map.jinja
which addresses all of the above issues.