Fix default packages
Previously the base option was not specified.  The default option only
specifies what dictionary to specify when the grain used for filtering
is not present.  The base option allows for the specified key to serve
as a key even if there does not exist a key in the matched dict.